aoc(day05): rust solution
After using rust for some time, I must say I really like the developer
ecosystem that has been created around it. With quite a lot of good
tool which is amazing.

But I find myself fighting against the language and more ways than one.
And keep missing zig. So for the rest of the AOC. I will be focusing
on Zig.

Maybe sometime in the future I might change my mind, but for now I want
to focus fully on zig. In my opinion, Zig is the systems programming
language of the future and it is already daring to think differently and
bring some innovative ideas in a lot of spaces.

Bye for now rust.